requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
numpy==1.19.2
Pillow==8.0.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.3
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.11
Werkzeug==1.0.1
